+++ source/DSCResources/MSFT_xDhcpServerOptionDefinition/MSFT_xDhcpServerOptionDefinition.psm1
@@ -89,6 +89,7 @@ function Get-TargetResource
             Type          = $dhcpServerOptionDefinition.Type
             VendorClass   = $dhcpServerOptionDefinition.VendorClass
             MultiValued   = $dhcpServerOptionDefinition.MultiValued
+            DefaultValue  = $dhcpServerOptionDefinition.DefaultValue
             Ensure        = 'Present'
         }
     }
@@ -102,6 +103,7 @@ function Get-TargetResource
             Type          = $null
             VendorClass   = $null
             MultiValued   = $false
+            DefaultValue  = $null
             Ensure        = 'Absent'
         }
     }
@@ -133,10 +135,13 @@ function Get-TargetResource
         The data type of the option definition.
 
     .PARAMETER MultiValued
-        Whether the option definition is multivalued or not.
+        Whether the option definition is multi-valued or not.
 
     .PARAMETER AddressFamily
         The option definition address family (IPv4 or IPv6). Currently only the IPv4 is supported.
+
+    .PARAMETER DefaultValue
+        Specifies the default value to set for the option definition.
 #>
 function Set-TargetResource
 {
@@ -181,7 +186,11 @@ function Set-TargetResource
         [Parameter(Mandatory = $true)]
         [ValidateSet('IPv4')]
         [System.String]
-        $AddressFamily
+        $AddressFamily,
+
+        [Parameter()]
+        [System.String]
+        $DefaultValue
     )
 
     # Reading the DHCP option
@@ -193,7 +202,7 @@ function Set-TargetResource
         # Testing if option exists
         if ($dhcpServerOptionDefinition.Ensure -eq 'Present')
         {
-            # If it exists and any of multivalued, type or vendorclass is being changed remove then re-add the whole option definition
+            # If it exists and any of multi-valued, type or vendor class is being changed remove then re-add the whole option definition
             if (($dhcpServerOptionDefinition.Type -ne $Type) -or ($dhcpServerOptionDefinition.MultiValued -ne $MultiValued) -or ($dhcpServerOptionDefinition.VendorClass -ne $VendorClass))
             {
                 $scopeIDMessage = $script:localizedData.RecreatingOptionDefinitionIDMessage -f $OptionId, $VendorClass
@@ -201,7 +210,21 @@ function Set-TargetResource
 
                 Remove-DhcpServerv4OptionDefinition -OptionId $OptionId -VendorClass $VendorClass
 
-                Add-DhcpServerv4OptionDefinition -OptionId $OptionId -name $Name -Type $Type -Description $Description -MultiValued:$MultiValued -VendorClass $VendorClass
+                $addDhcpServerv4OptionDefinitionParameters = @{
+                    OptionId = $OptionId
+                    Name = $Name
+                    Type = $Type
+                    Description = $Description
+                    MultiValued = $MultiValued
+                    VendorClass = $VendorClass
+                }
+
+                if ($PSBoundParameters.ContainsKey('DefaultValue'))
+                {
+                    $addDhcpServerv4OptionDefinitionParameters.DefaultValue = $DefaultValue
+                }
+
+                Add-DhcpServerv4OptionDefinition @addDhcpServerv4OptionDefinitionParameters
             }
             # If option exists we need only to adjust the parameters
             else
@@ -209,7 +232,19 @@ function Set-TargetResource
                 $settingIDMessage = $script:localizedData.SettingOptionDefinitionIDMessage -f $OptionId, $VendorClass
                 Write-Verbose -Message $settingIDMessage
 
-                Set-DhcpServerv4OptionDefinition -OptionId $OptionId -VendorClass $VendorClass -name $Name -Description $Description
+                $setDhcpServerv4OptionDefinitionParameters = @{
+                    OptionId = $OptionId
+                    Name = $Name
+                    Description = $Description
+                    VendorClass = $VendorClass
+                }
+
+                if ($PSBoundParameters.ContainsKey('DefaultValue'))
+                {
+                    $setDhcpServerv4OptionDefinitionParameters.DefaultValue = $DefaultValue
+                }
+
+                Set-DhcpServerv4OptionDefinition @setDhcpServerv4OptionDefinitionParameters
             }
         }
         # If option does not exist we need to add it
@@ -218,7 +253,21 @@ function Set-TargetResource
             $scopeIDMessage = $script:localizedData.AddingOptionDefinitionIDMessage -f $OptionId, $VendorClass
             Write-Verbose -Message $scopeIDMessage
 
-            Add-DhcpServerv4OptionDefinition -OptionId $OptionId -name $Name -Type $Type -Description $Description -MultiValued:$MultiValued -VendorClass $VendorClass
+            $addDhcpServerv4OptionDefinitionParameters = @{
+                OptionId = $OptionId
+                Name = $Name
+                Type = $Type
+                Description = $Description
+                MultiValued = $MultiValued
+                VendorClass = $VendorClass
+            }
+
+            if ($PSBoundParameters.ContainsKey('DefaultValue'))
+            {
+                $addDhcpServerv4OptionDefinitionParameters.DefaultValue = $DefaultValue
+            }
+
+            Add-DhcpServerv4OptionDefinition @addDhcpServerv4OptionDefinitionParameters
         }
     }
     # Testing for 'absent'
@@ -262,6 +311,9 @@ function Set-TargetResource
 
     .PARAMETER AddressFamily
         The option definition address family (IPv4 or IPv6). Currently only the IPv4 is supported.
+
+    .PARAMETER DefaultValue
+        Specifies the default value to set for the option definition.
 #>
 function Test-TargetResource
 {
@@ -307,7 +359,11 @@ function Test-TargetResource
         [Parameter(Mandatory = $true)]
         [ValidateSet('IPv4')]
         [System.String]
-        $AddressFamily
+        $AddressFamily,
+
+        [Parameter()]
+        [System.String]
+        $DefaultValue
     )
 
     # Region Input Validation
@@ -349,7 +405,7 @@ function Test-TargetResource
                 of the parameters need to be evaluated if they that they are in desired
                 state.
             #>
-            $propertiesToEvaluate = @('Name','Description','Type','MultiValued')
+            $propertiesToEvaluate = @('Name', 'Description', 'Type', 'MultiValued', 'DefaultValue')
 
             $result = $true
 
@@ -360,9 +416,20 @@ function Test-TargetResource
                 {
                     $desiredParameterValue = Get-Variable -Name $property -ValueOnly
 
-                    if ($currentConfiguration.$property -ne $desiredParameterValue)
+                    if ($property -eq 'DefaultValue')
                     {
-                        $result = $false
+                        # Force string comparison, else get mixed results with DefaultValue property
+                        if ([System.String] $currentConfiguration.$property -ne [System.String] $desiredParameterValue)
+                        {
+                            $result = $false
+                        }
+                    }
+                    else
+                    {
+                        if ($currentConfiguration.$property -ne $desiredParameterValue)
+                        {
+                            $result = $false
+                        }
                     }
                 }
             }